% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_design_matrix}
\alias{sim_design_matrix}
\title{Design Matrix}
\usage{
sim_design_matrix(n, nx = 2, add_intercept = TRUE)
}
\arguments{
\item{n}{integer. Number of observations.}

\item{nx}{integer. Number of explanatory variables, excluding intercept.}

\item{add_intercept}{logical. An intercept column is prepened when \code{TRUE}.}
}
\description{
Simulated from i.i.d N(0,1) distributions, optional intercept.
}
