% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_group_membership}
\alias{sim_group_membership}
\title{Group Membership}
\usage{
sim_group_membership(n, prob)
}
\arguments{
\item{n}{integer. Number of observations.}

\item{prob}{passed to \code{\link{sample}}.}
}
\description{
Returns (factor of) groupings based on a vector of probabilities.
}
