% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_mixregr}
\alias{sim_mixregr}
\title{Mixture of Regressions}
\usage{
sim_mixregr(x, z, regr, sd)
}
\arguments{
\item{x}{matrix. Design matrix.}

\item{z}{integer. Group membership variable.}

\item{regr}{matrix. Regression matrix with \code{p} rows and \code{k} columns.}

\item{sd}{numeric. Standard deviation of the observational error.}
}
\description{
Simulate from \code{K} (\code{length(groups)}) distinct regressions based on group
membership variable \code{z}.
}
\details{
That is,
\eqn{Y_i \sim N(x_i * regr_{i, z_i}, sd^2)}
}
