% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{vis}
\alias{vis}
\alias{vis_scatterplot}
\alias{vis_boxgroup}
\title{Visualisation Methods}
\usage{
vis_scatterplot(x, y, z)

vis_boxgroup(y, z)
}
\arguments{
\item{x}{matrix. Design matrix}

\item{y}{vector. Response variable data.}

\item{z}{vector/factor. Group membership data.}
}
\description{
Visualisation Methods
}
\section{Functions}{
\itemize{
\item \code{vis_scatterplot}: Scatterplot with additional group membership data shown as colour.

\item \code{vis_boxgroup}: Boxplot of \code{y} for each group (independent of \code{x}).
}}

